Fixed timer bug- timer speed is constant Moved panels to a separate folder
