+++ src/view/panels/SettingsPanel.java
@@ -1,8 +1,9 @@
-package view;
+package view.panels;
+
+import view.Constants;
 
 import javax.swing.*;
 import java.awt.*;
-import java.awt.event.ActionEvent;
 
 
 public class SettingsPanel extends MyPanel {
